Move to shared - ValueUtilsSmi (#1326)
